backend/requirements.txt
# Flask Framework
Flask==3.0.0
Flask-CORS==4.0.0

# MongoDB
pymongo==4.6.1

# Authentication & Security
PyJWT==2.8.0
bcrypt==4.1.2
python-dotenv==1.0.0

# Image Processing
Pillow==10.1.0
opencv-python==4.8.1.78

# Machine Learning (from existing project)
tensorflow==2.15.0
numpy==1.24.3

# Utilities
Werkzeug==3.1.4

# Timezone Support
pytz==2023.3

# Excel Export
openpyxl==3.1.2

# Production Servers
waitress==2.1.2  # For Windows
gunicorn==21.2.0  # For Linux

# Testing
pytest==7.4.3
pytest-flask==1.3.0
